Fix #83, MailSummary fails in production.
Fixes error where mailer sender/from field was not being set for mail
messages that should not have been sent.  Unfortunately, the logic to
determine whether or not to send the mail is more easily put in the
mailer controller for now, even though it should probably not go there.